Long & defs

+++ file_generator.c
@@ -18,7 +18,7 @@ void test_file(FILE *fp)
 
 // Item item_constructor(int data, int key);
 
-int rand_num_by_max(int max)
+int rand_num_by_max(long int max)
 {
     return 1 + rand() % max;
 }
@@ -38,7 +38,7 @@ void create_string(int max_size, char str[])
     str[size - 1] = '\0';
 }
 
-void build_file(int nPages, int page_size, int data_max, int key_dif) // creates a file with random data
+void build_file(int nPages, int page_size, long int data_max, int key_dif) // creates a file with random data
 {
     FILE *fp = fopen("data.bin", "wb");
     test_file(fp);
